[2025][day07] Add scala solution for part 2
Solution with dynamic programming.
The recursive breakdow is:
number of timelines a ray can reach the bottom from the space at coordinate (i, j) is:
1. the same as the solution for the space right under it at coordinate (i+1, j).
2. the sum of the solution to the two spaces:
    * under it to the left at coordinate (i+1. j-1)
    * under it to the left at coordinate (i+1. j+1)
Bound checking is needed to make sure we don't access invalid spaces.

Start calculating from bottom to top (decreassing i) because knowing the answer for any coordinate (i, j) requires
knowing the answer for the row below it.